vis-lua: replace all unnecessary string with length operations
lua uses strings with length so we should use that to our
advantage. Also while we are at it many of the strings we push are
literals so we should use lua_pushliteral instead of
lua_push?string. When we do know the length we should use
lua_pushlstring (most cases). The remaining non-lengthed strings
will likely be removed as more of the codebase adapts s8s