replace oversized libutf with smaller version
this is taken from one of my other projects. there was no reason
for there to be 2x the code

tests checking for surrogate characters and non characters were
removed. I see no reason why the user shouldn't be allowed to
insert those characters in text (they exist in the standard).
Also, in the case of non-characters only the first two were being
checked and not the other 64.