util: replace memrchr with internal version
The amount of code we need to detect if this is present and handle
the fallback is more than if we just provide it ourselves.

The previously provided fallback misbehaves if 'n' is 0, which is
probably brain damage allowed by the standard, so I wrote a better
version. Also we are passing in a difference of pointers, which is
a very common usage for any of those memory functions, so the
argument type should be ptrdiff_t. This avoids another C brain
damage of having unsigned size type.